feat(timebase): per-mode ranges — Rolling to 1 s/div, XY to 100 ms/div
The three modes no longer share one timebase.  Now that the free-running
frame rate is bounded by the blit rather than the sample rate, Rolling and
XY can usefully go much slower than Triggered:

  Triggered : 50 µs … 10 ms/div  (unchanged)
  Rolling   : 500 µs … 1 s/div
  XY        : 500 µs … 100 ms/div

Because the ranges differ, each mode remembers its own timebase and a mode
switch restores it (setTimebase()/timebase() index a per-mode array).

- ScopeState: timebase_us_per_div becomes a uint32 array indexed by Mode.
  uint32 is required — 1 s = 1,000,000 µs overflows the old uint16.  EEPROM
  StoredState version bumped 2 -> 3 (old records reset to defaults on load).
- Parameter: per-mode step tables (shared 1-1.5-2-3-5-7 sequence), and
  fmtTime gains a seconds branch.
- Acquisition: timebase widened to uint32 throughout; the free-running path
  now draws a partial window while the first full one fills, so 1 s/div
  (~30 Hz, ~8 s to fill 240 px) grows in from the left instead of showing a
  blank screen.  CellStats.timebase widened to match.

Verified on hardware: 1 s/div Rolling holds 30 fps (sample-bound, smooth
1 px/frame scroll); 100 ms/div XY holds ~49 fps (blit-bound); zero tears,
overruns, or A/B skew at both extremes.  The ~30 Hz sample timer drives the
eDMA rings correctly.

+++ lib/AcqCore/AcqCore.h
@@ -61,7 +61,7 @@ inline bool newestWindow(uint64_t safe, uint32_t len, uint64_t* base) {
 // and worst-second peaks so the reporter can print one line per cell that maps
 // straight onto the protocol table (docs/acq-characterization.md).
 struct CellStats {
-    uint16_t timebase = 0;      // µs/div this cell is for
+    uint32_t timebase = 0;      // µs/div this cell is for
     uint8_t  mode     = 0xFF;   // ScopeState Mode, as a plain byte (no Arduino dep)
 
     uint32_t secs     = 0;
@@ -81,7 +81,7 @@ struct CellStats {
 
     // Begin a fresh cell.  Called whenever the timebase or mode changes, so a
     // knob sweep through intermediate steps never pollutes a dwell.
-    void restart(uint16_t tb, uint8_t m) {
+    void restart(uint32_t tb, uint8_t m) {
         *this = CellStats{};
         timebase = tb;
         mode     = m;
